first version of the new ANTLR based TFTASolver approach
the core of this new TTASolver is based on ANTRL grammars
they provide Java code and result in a stand alone Java program called
tftac ("TFTA Compiler").

The OpenPascal GUI program tftasolver is kept but stripped of all the old
"internals", namely the tftalogic, tftaexpression, tftaobject units.

The new tftac is capable of compiling TFTA terms but has no support of
syntactic analysis, i.e. the compiler works (although not thoroughly tested
yet) but can not understand the "values" of the events. As such the temporal
absorptions laws and similar other laws are not implemented yet, there is no
possibility of generating a minimal form or a disjunct form and so on...

Signed-off-by: sjschilling <sjschilling@users.berlios.de>